Convert Scene project to EnhancedInput
Summary:
NOTE: As of 5.4, unreal will remove the support for Legacy inputs. This diff converts old inputs to the new Enhanced Input system.

Due to the simplicity of the interactions in this project, only one Input Mapping Context is needed and was implemented.

Also removed the `Controllers/` folder and all assets underneath it, it was only used in the VRCharacter BP but hidden in game so not impacting the sample in anyway.

Reviewed By: MartinSherburn

Differential Revision: D51027327

fbshipit-source-id: 511245bb78e94967fc4c0f98846f1dd8e4193b7a